exercicios

+++ week-03/expressoes-booleanas/impar-ou-par.py
@@ -1,7 +1,5 @@
-numero = (float(input("Digite um número: ")))
-
+numero = (int(input("Digite um número: ")))
 result = numero / 2
-
 if result == int(result):
     print ("O número é par!")
 else:
